functions.cpp translation unit
