+++ CMakeLists.txt
@@ -6,6 +6,7 @@ include(CTest)
 set(CMAKE_CXX_STANDARD 20)
 
 add_executable(Multiblend
+  "src/functions.cpp"
   "src/multiblend.cpp"
 )
